+++ src/cgame/cg_draw.cpp
@@ -220,19 +220,19 @@ static void CG_DrawBeacon( cbeacon_t *b )
 	                       CG_BeaconIcon( b ) );
 
 	if( b->flags & EF_BC_DYING )
-		trap_R_DrawStretchPic( b->pos[ 0 ] - b->size/2 * 1.3,
-		                       b->pos[ 1 ] - b->size/2 * 1.3,
-		                       b->size * 1.3, b->size * 1.3,
+		trap_R_DrawStretchPic( b->pos[ 0 ] - b->size/2 * 1.3f,
+		                       b->pos[ 1 ] - b->size/2 * 1.3f,
+		                       b->size * 1.3f, b->size * 1.3f,
 		                       0, 0, 1, 1,
 		                       cgs.media.beaconNoTarget );
 
 	if ( b->clamped )
-		trap_R_DrawRotatedPic( b->pos[ 0 ] - b->size/2 * 1.5,
-		                       b->pos[ 1 ] - b->size/2 * 1.5,
-		                       b->size * 1.5, b->size * 1.5,
+		trap_R_DrawRotatedPic( b->pos[ 0 ] - b->size/2 * 1.5f,
+		                       b->pos[ 1 ] - b->size/2 * 1.5f,
+		                       b->size * 1.5f, b->size * 1.5f,
 		                       0, 0, 1, 1,
 		                       cgs.media.beaconIconArrow,
-		                       270.0 - ( angle = atan2( b->clamp_dir[ 1 ], b->clamp_dir[ 0 ] ) ) * 180 / M_PI );
+		                       270.0f - ( angle = atan2f( b->clamp_dir[ 1 ], b->clamp_dir[ 0 ] ) ) * 180 / M_PI );
 
 	if( b->type == BCT_TIMER )
 	{
@@ -247,7 +247,7 @@ static void CG_DrawBeacon( cbeacon_t *b )
 			vec2_t pos, dir, rect[ 2 ];
 			int i, l, frame;
 
-			h = b->size * 0.4;
+			h = b->size * 0.4f;
 			p = va( "%d", num );
 			l = strlen( p );
 			tw = h * l;
@@ -538,12 +538,12 @@ static void CG_DrawBinaryShadersFinalPhases()
 		return;
 	}
 
-	ss = sqrt( ss );
+	ss = sqrtf( ss );
 
 	trap_Cvar_VariableStringBuffer( "r_znear", str, sizeof( str ) );
 	f = atof( str ) + 0.01;
-	l = f * tan( DEG2RAD( cg.refdef.fov_x / 2 ) ) * ss;
-	u = f * tan( DEG2RAD( cg.refdef.fov_y / 2 ) ) * ss;
+	l = f * tanf( DEG2RAD( cg.refdef.fov_x / 2 ) ) * ss;
+	u = f * tanf( DEG2RAD( cg.refdef.fov_y / 2 ) ) * ss;
 
 	VectorMA( cg.refdef.vieworg, f, cg.refdef.viewaxis[ 0 ], verts[ 0 ].xyz );
 	VectorMA( verts[ 0 ].xyz, l, cg.refdef.viewaxis[ 1 ], verts[ 0 ].xyz );
