+++ src/cgame/rocket/rocketDataSource.h
@@ -232,8 +232,8 @@ protected:
 			child = GetChild( i );
 			width = child->GetOffsetWidth();
 			height = child->GetOffsetHeight();
-			float y = sin( angle * ( i - 1 ) * ( M_PI / 180.0f ) ) * radius;
-			float x = cos( angle * ( i - 1 ) * ( M_PI / 180.0f ) ) * radius;
+			float y = sinf( angle * ( i - 1 ) * ( M_PI / 180.0f ) ) * radius;
+			float x = cosf( angle * ( i - 1 ) * ( M_PI / 180.0f ) ) * radius;
 
 			child->SetProperty( "position", "absolute" );
 			child->SetProperty( "left", va( "%fpx", ( dimensions.x / 2 ) - ( width / 2 ) + offset.x + x ) );
